+++ Sources/KeyboardSupport/KeyboardRespondable.swift
@@ -54,12 +54,12 @@ extension UIViewController {
 
 /// Stores info about the keyboard.
 public struct KeyboardInfo {
-    let initialFrame: CGRect
-    let finalFrame: CGRect
-    let animationDuration: TimeInterval
-    let animationCurve: UInt
+    public let initialFrame: CGRect
+    public let finalFrame: CGRect
+    public let animationDuration: TimeInterval
+    public let animationCurve: UInt
     
-    init?(notification: Notification) {
+    public init?(notification: Notification) {
         #if swift(>=4.2)
         guard let userInfo = notification.userInfo,
             let initialKeyboardFrame = userInfo[UIResponder.keyboardFrameBeginUserInfoKey] as? CGRect,
@@ -84,7 +84,7 @@ public struct KeyboardInfo {
         animationCurve = curve
     }
     
-    var isMoving: Bool {
+    public var isMoving: Bool {
         return initialFrame.origin != finalFrame.origin
     }
 }
@@ -112,16 +112,6 @@ public protocol KeyboardScrollable: class {
     
     /// Called when the keyboard is hiding.
     func keyboardWillHide(keyboardInfo: KeyboardInfo)
-}
-
-public extension KeyboardScrollable {
-    func keyboardWillShow(keyboardInfo: KeyboardInfo) {
-        // No-op by default. Opt-in by implementing this method in your class conforming to KeyboardScrollable.
-    }
-    
-    func keyboardWillHide(keyboardInfo: KeyboardInfo) {
-        // No-op by default. Opt-in by implementing this method in your class conforming to KeyboardScrollable.
-    }
 }
 
 public extension KeyboardScrollable where Self: UIViewController {
@@ -173,11 +163,11 @@ public extension KeyboardScrollable where Self: UIViewController {
         }
     }
     
-    func keyboardWillShow() {
+    func keyboardWillShow(keyboardInfo: KeyboardInfo) {
         // No-op by default. Opt-in by implementing this method in your class conforming to KeyboardScrollable.
     }
     
-    func keyboardWillHide() {
+    func keyboardWillHide(keyboardInfo: KeyboardInfo) {
         // No-op by default. Opt-in by implementing this method in your class conforming to KeyboardScrollable.
     }